Added Exception Handling for Service Object Serive

+++ src/main/java/com/swisscom/tasks/task3/service/impl/ServiceOServiceImpl.java
@@ -1,5 +1,6 @@
 package com.swisscom.tasks.task3.service.impl;
 
+import com.swisscom.tasks.task3.exception.ServiceOServiceException;
 import com.swisscom.tasks.task3.model.ServiceO;
 import com.swisscom.tasks.task3.repository.ServiceORepository;
 import com.swisscom.tasks.task3.service.ServiceOService;
@@ -16,8 +17,10 @@ public class ServiceOServiceImpl implements ServiceOService {
 
     private final ServiceORepository serviceORepository;
     @Override
-    public String save(ServiceO serviceO) {
-        return serviceORepository.save(serviceO).getId();
+    public ServiceO save(ServiceO serviceO) {
+        if(serviceO.getId() !=null && serviceORepository.existsById(serviceO.getId()))
+            throw new ServiceOServiceException("Another service with this ID exists before");
+        return serviceORepository.save(serviceO);
     }
 
     @Override
@@ -32,6 +35,9 @@ public class ServiceOServiceImpl implements ServiceOService {
 
     @Override
     public boolean deleteById(String id) {
+        boolean exists=serviceORepository.existsById(id);
+        if(!exists)
+            throw new ServiceOServiceException("Service with this ID does not exist");
         serviceORepository.deleteById(id);
         return true;
     }
